% Generated by build script
\name{otopo_catalog}
\alias{otopo_catalog}
\title{Search OpenTopography dataset catalog by bounding box}
\usage{
otopo_catalog()
}
\arguments{
\item{minx}{Minimum longitude (e.g. -118)}
\item{miny}{Minimum latitude (e.g. 33)}
\item{maxx}{Maximum longitude (e.g. -117)}
\item{maxy}{Maximum latitude (e.g. 34)}
\item{product_format}{Product format filter: "PointCloud" or "Raster"}
\item{detail}{Include detailed metadata (default TRUE)}
}
\value{tibble: name, short_name, dataset_type, platform, minx, miny,}
\description{Search OpenTopography dataset catalog by bounding box}
