% Generated by build script
\name{otopo_global_datasets}
\alias{otopo_global_datasets}
\title{List all global datasets available on OpenTopography}
\usage{
otopo_global_datasets()
}
\arguments{
\item{product_format}{"PointCloud" or "Raster" (default "PointCloud")}
}
\value{tibble: name, short_name, dataset_type, platform, minx, miny,}
\description{List all global datasets available on OpenTopography}
