% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otopo.R
\name{otopo_catalog}
\alias{otopo_catalog}
\title{Search OpenTopography dataset catalog by bounding box}
\usage{
otopo_catalog(
  minx,
  miny,
  maxx,
  maxy,
  product_format = "PointCloud",
  detail = TRUE
)
}
\arguments{
\item{minx}{Minimum longitude (e.g. -118)}

\item{miny}{Minimum latitude (e.g. 33)}

\item{maxx}{Maximum longitude (e.g. -117)}

\item{maxy}{Maximum latitude (e.g. 34)}

\item{product_format}{Product format filter: "PointCloud" or "Raster"
(default "PointCloud")}

\item{detail}{Include detailed metadata (default TRUE)}
}
\value{
tibble: name, short_name, dataset_type, platform, minx, miny,
maxx, maxy, url
}
\description{
Search OpenTopography dataset catalog by bounding box
}
